+++ smartmemory/code/__init__.py
@@ -3,5 +3,14 @@
 from smartmemory.code.models import CodeEntity, CodeRelation, ParseResult, IndexResult
 from smartmemory.code.parser import CodeParser
 from smartmemory.code.indexer import CodeIndexer
+from smartmemory.code.search import semantic_code_search
 
-__all__ = ["CodeEntity", "CodeRelation", "ParseResult", "IndexResult", "CodeParser", "CodeIndexer"]
+__all__ = [
+    "CodeEntity",
+    "CodeRelation",
+    "ParseResult",
+    "IndexResult",
+    "CodeParser",
+    "CodeIndexer",
+    "semantic_code_search",
+]